refactor: shared/api 공개 API 경계 정리, 에러 파싱 중복 제거
코드리뷰 지적사항 반영:
- entities/member, features/login, features/screening-flow가
  shared/api/http를 직접 import하던 것을 shared/api(index.ts)를
  통하도록 수정. ADR-0003의 "밖에는 api.*만 노출한다" 결정과
  어긋났던 부분이라 ADR-0004에 후속 결정으로 이유를 남김
- 생성 클라이언트가 던지는 {error: ApiErrorResponse} 형태를 각자
  다르게 캐스팅해 풀던 걸 shared/api/parse-api-error.ts로 통합
