+++ src/features/login/model/login-with-kakao-code.ts
@@ -1,6 +1,4 @@
-import { setAccessToken } from "@/shared/api";
-import { authApi } from "@/shared/api/http";
-import type { ApiErrorResponse } from "@/shared/api/generated/data-contracts";
+import { authApi, parseApiError, setAccessToken } from "@/shared/api";
 
 export type LoginWithKakaoCodeResult =
   { success: true } | { success: false; code: string; message: string };
@@ -17,7 +15,7 @@ export async function loginWithKakaoCode(
     setAccessToken(accessToken);
     return { success: true };
   } catch (err) {
-    const apiError = (err as { error?: ApiErrorResponse }).error;
+    const apiError = parseApiError(err);
     return {
       success: false,
       code: apiError?.code ?? "UNKNOWN",
