+++ src/shared/api/index.ts
@@ -1,2 +1,4 @@
 export { api } from "./api";
 export { isAuthenticated, setAccessToken } from "./access-token";
+export { authApi, coursesApi, membersApi, screeningApi } from "./http";
+export { parseApiError } from "./parse-api-error";
